+++ src/message/mod.rs
@@ -6,9 +6,11 @@ pub mod message_type;
 pub mod serialized_message;
 pub mod system_stop_message;
 pub mod delayed_message;
+pub mod sleep_message;
 
 pub mod prelude {
     pub use crate::message::actor_message::ActorMessage;
     pub use crate::message::bulk_actor_message::BulkActorMessage;
     pub use crate::message::serialized_message::SerializedMessage;
+
 }